requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.6
h11==0.14.0
idna==3.10
numpy==1.24.4
opencv-python-headless==4.7.0.72
packaging==24.2
pillow==11.1.0
pydantic==2.10.5
pydantic_core==2.27.2
pytesseract==0.3.13
RPi.GPIO==0.7.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
